requirements.txt
Flask==2.3.2
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.1
google-api-python-client==2.91.0
python-dotenv==1.0.0
Werkzeug==2.3.6
